renames projects container to overview
