+++ client/src/containers/overview/header/index.tsx
@@ -10,7 +10,7 @@ import { projectsUIState } from "@/app/(overview)/store";
 import { useGlobalFilters } from "@/app/(overview)/url-store";
 import { INITIAL_FILTERS_STATE } from "@/app/(overview)/url-store";
 
-import ParametersProjects from "@/containers/projects/header/parameters";
+import ParametersProjects from "@/containers/overview/header/parameters";
 
 import { Button } from "@/components/ui/button";
 import { SidebarTrigger } from "@/components/ui/sidebar";
