+++ client/src/containers/overview/map/controls/zoom/index.tsx
