+++ client/src/containers/overview/table/view/scorecard-prioritization/columns.tsx
